Format and sort code correctly
