+++ lib/pangea/activity_planner/activity_plan_message.dart
@@ -1,7 +1,7 @@
 import 'package:fluffychat/config/themes.dart';
 import 'package:fluffychat/pages/chat/chat.dart';
 import 'package:fluffychat/pages/chat/events/message_content.dart';
-import 'package:fluffychat/pages/chat/events/message_reactions.dart';
+import 'package:fluffychat/pages/chat/events/pangea_message_reactions.dart';
 import 'package:fluffychat/utils/date_time_extension.dart';
 import 'package:fluffychat/widgets/matrix.dart';
 import 'package:flutter/material.dart';
@@ -198,8 +198,11 @@ class ActivityPlanMessage extends StatelessWidget {
                                 top: 4.0,
                                 right: 4.0,
                               ),
-                              child:
-                                  MessageReactions(event, timeline, controller),
+                              child: PangeaMessageReactions(
+                                event,
+                                timeline,
+                                controller,
+                              ),
                             ),
                           ],
                         ),
